Minor fixes
Added a pause-menu that is activated by escape.
The inventory can now be closed by pressing the same button as the
number of the item that is showed.
Fixed a bug in the time-counter.
Repositioned Fanton's waypoints, some of them was outside the map and
some really close to each other.
Fixed a bug where the mouse would appear when it shouldn't.
The enemies can't cause damage to Fanton when they are walking towards
tequila no more.